Add files via upload
The "MathUtils" object contains the factorial method. This method calculates the factorial of a given number using recursion. If the number is 0 or 1, it returns 1. Otherwise, it recursively calls itself with n - 1 and multiplies the result by n. The "Main" object contains the main method where you can test the factorial method. In this example, it calculates the factorial of the number 4 and 10 and prints the result.